incremental GC fcntl patch (16-Nov-96)

+++ m3/pm3/libs/m3core/src/runtime/ALPHA_OSF/RTHeapDepC.c
@@ -410,7 +410,7 @@ int exportfs(access, cookie, exdata)
 
 int fcntl(filedes, request, argument)
   int filedes, request;
-  int argument;
+  long argument;
 {   int result;
 
     ENTER_CRITICAL;
